Feature/authorization

+++ nourishify.api/Shared/Infrastructure/Persistence/Repositories/BaseRepository.cs
@@ -18,7 +18,7 @@ public class BaseRepository<TEntity> : IBaseRepository<TEntity> where TEntity : 
         await Context.Set<TEntity>().AddAsync(entity);
     }
 
-    public async Task<TEntity?> FindByIdAsync(int id)
+    public async Task<TEntity?> FindByIdAsync(long id)
     {
         return await Context.Set<TEntity>().FindAsync(id);
     }
